задание 5.9
