+++ src/view/sorting.js
@@ -1,14 +1,14 @@
 import AbstractView from '../framework/view/abstract-view.js';
-export const Sort = {
-  PRICE:'sort-price',
+export const SortType = {
+  PRICE: 'sort-price',
   DAY: 'sort-day',
   TIME: 'sort-time'
 };
 function getSortingTemplate() {
   return `<form class="trip-events__trip-sort  trip-sort" action="#" method="get">
             <div class="trip-sort__item  trip-sort__item--day">
-              <input id="sort-day" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-day" checked="" data-id="${Sort.day}">
-              <label class="trip-sort__btn" for="${Sort.DAY}">Day</label>
+              <input id="sort-day" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-day" checked="" data-id="${SortType.day}">
+              <label class="trip-sort__btn" for="${SortType.DAY}">Day</label>
             </div>
 
             <div class="trip-sort__item  trip-sort__item--event">
@@ -17,13 +17,13 @@ function getSortingTemplate() {
             </div>
 
             <div class="trip-sort__item  trip-sort__item--time">
-              <input id="sort-time" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-time" data-id="${Sort.time}">
-              <label class="trip-sort__btn" for="${Sort.TIME}">Time</label>
+              <input id="sort-time" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-time" data-id="${SortType.time}">
+              <label class="trip-sort__btn" for="${SortType.TIME}">Time</label>
             </div>
 
             <div class="trip-sort__item  trip-sort__item--price">
-              <input id="sort-price" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-price" data-id="${Sort.price}">
-              <label class="trip-sort__btn" for="${Sort.PRICE}">Price</label>
+              <input id="sort-price" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-price" data-id="${SortType.price}">
+              <label class="trip-sort__btn" for="${SortType.PRICE}">Price</label>
             </div>
 
             <div class="trip-sort__item  trip-sort__item--offer">
